⚡️ Add Dynamic Priority Fee (#6)
### 🕓 Changelog

This PR adds an example how to dynamically calculate the priority fee
when constructing a transaction. Specifically, the following logic is
added:
- The current base fee is fetched from the latest block.
- A dynamic buffer decrease of **0.5%** is applied to account for
potential base fee increases in the next block. It's important to note
that the base fee can increase by a maximum of 12.5% in the next block.
- The priority gas price (tip) is then calculated by subtracting the
base fee and buffer from the maximum fee per gas (i.e., `GAS_PRICE`).
- If the calculated priority gas price is negative (e.g. due to base fee
query errors on the RPC side), it is set it to **5 gwei**.

---------

Signed-off-by: Pascal Marco Caversaccio <pascal.caversaccio@hotmail.ch>